move scripts to dedicated forlder
