+++ ZadanieWav/utils.py
